+++ client/src/main/java/net/oneandone/stool/client/Client.java
@@ -192,7 +192,7 @@ public class Client {
         JsonObject response;
         Map<String, Map<String, String>> result;
 
-        response = getJson(node(stage, "await-startup").withParameter("legacy", "false")).getAsJsonObject();
+        response = getJson(node(stage, "await-startup")).getAsJsonObject();
         result = new LinkedHashMap<>();
         for (Map.Entry<String, JsonElement> entry : response.entrySet()) {
             result.put(entry.getKey(), stringMap(entry.getValue().getAsJsonObject()));